requirements.txt
# Icon conversion
cairosvg==2.7.1
pillow==10.4.0